--- src/Breadcrumb/Breadcrumb.md
@@ -1,12 +1,0 @@
-```jsx
-<div style={{ display: "flex" }}>
-	<Breadcrumb.List>
-		<Breadcrumb.Item text={"Призы и подарки"} />
-		<Breadcrumb.Item text={"Хлопушки, пряники, конфетти"} />
-		<Breadcrumb.Item text={"Акции для постоянных клиентов"} />
-		<Breadcrumb.Item text={"Подарочные акции"} />
-		<Breadcrumb.Item text={"2019-2020"} />
-		<Breadcrumb.Item text={"Снова в школу 2019"} />
-	</Breadcrumb.List>
-</div>
-```
