--- src/Breadcrumb/index.ts
@@ -1,1 +1,0 @@
-export * from "./Breadcrumb";
