--- src/Breadcrumb/styles/List.scss
@@ -1,26 +1,0 @@
-.kit-breadcrumb-list {
-	position: relative;
-	margin: 0;
-	padding: 21px 30px 11px 5px;
-	display: flex;
-	align-items: center;
-	justify-content: flex-start;
-	flex-wrap: wrap;
-	width: 100%;
-	background-color: #f3f3eb;
-	border-radius: 4px;
-	border: 1px dashed #d4d4d4;
-	box-shadow: 5px 5px 0 0 #e6e6dd;
-
-	&::after {
-		content: "";
-		position: absolute;
-		top: 5px;
-		left: 5px;
-		height: 100%;
-		width: 100%;
-		border-bottom: 1px dashed #d4d4d4;
-		border-right: 1px dashed #d4d4d4;
-		border-radius: 4px;
-	}
-}
