--- src/Breadcrumb/styles/index.scss
@@ -1,2 +1,0 @@
-@import "List";
-@import "Item";
