+++ src/Breadcrumbs/Breadcrumbs.md
@@ -1,0 +1,12 @@
+```jsx
+<div style={{ display: "flex" }}>
+	<Breadcrumbs>
+		<Breadcrumbs.Item text={"Призы и подарки"} />
+		<Breadcrumbs.Item text={"Хлопушки, пряники, конфетти"} />
+		<Breadcrumbs.Item text={"Акции для постоянных клиентов"} />
+		<Breadcrumbs.Item text={"Подарочные акции"} />
+		<Breadcrumbs.Item text={"2019-2020"} />
+		<Breadcrumbs.Item text={"Снова в школу 2019"} />
+	</Breadcrumbs>
+</div>
+```
